Namespacing Job and Serializer Classes
