--- app/serializers/authorization_serializer.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-class AuthorizationSerializer
-  include FastJsonapi::ObjectSerializer
-  attributes :id, :code
-end
