+++ app/serializers/browse_everything/authorization_serializer.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module BrowseEverything
+  class AuthorizationSerializer
+    include FastJsonapi::ObjectSerializer
+    attributes :id, :code
+  end
+end
