+++ app/serializers/browse_everything/bytestream_serializer.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module BrowseEverything
+  class BytestreamSerializer
+    include FastJsonapi::ObjectSerializer
+    attributes :location, :name, :size, :mtime, :media_type, :uri
+  end
+end
