+++ app/serializers/browse_everything/container_serializer.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module BrowseEverything
+  class ContainerSerializer
+    include FastJsonapi::ObjectSerializer
+    attributes :id, :name, :mtime
+
+    has_many :bytestreams
+    has_many :containers
+  end
+end
