+++ app/serializers/browse_everything/upload_serializer.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module BrowseEverything
+  class UploadSerializer
+    include FastJsonapi::ObjectSerializer
+    attributes :session_id, :bytestream_ids, :container_ids
+
+    has_one :session
+    has_many :bytestreams
+    has_many :containers
+  end
+end
