--- app/serializers/provider_serializer.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-class ProviderSerializer
-  include FastJsonapi::ObjectSerializer
-  attributes :id, :name
-
-  # rubocop:disable Style/SymbolProc
-  link :authorization_url do |object|
-    object.authorization_url
-  end
-  # rubocop:enable Style/SymbolProc
-end
